Boundary Perturbation Inflator:     Simplified periodic dof assignment in boundary perturbation inflator, also     fixing a bug in the 3D case.
    Fixed the normalization of the free boundary gradients on the periodic
    boundary (was causing bad behavior at the "connectors").

    Implemented conversion of parameter vector to corresponding boundary vector
    field for the boundary perturbation inflator.

Boundary Curve Remesher:
    Fixed bug where certain collapse operations caused a non-closed boundary curve
    (resulting in an empty mesh from Triangle).

    Improved handling of feature vertices in the collapse criteria (ensures
    nearby feature vertices are still merged).

WCS Optimization
    Normalized all sub-objectives, so weights are more consistent across
    meshes.

    Added output of sub-objective shape derivatives for debugging

New executable WCS_cli: stand-alone worst case stress computation.

+++ pattern_optimization/PatternOptimization.hh
@@ -341,7 +341,7 @@ public:
             size_t r = 0;
             for (size_t i = 0; i < flatLen(_N); ++i) {
                 for (size_t j = i; j < flatLen(_N); ++j) {
-                    for (size_t p = 0; p < numParams; ++p)
+                    for (size_t p = 0; p < size_t(numParams); ++p)
                         jacobian[r * numParams + p] = m_iterate->jacobian(i, j, p);
                     ++r;
                 }
